+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Automatically migrate legacy config files to the new format (Will be removed in 1.9.0)
 - Add more comments to better clarify things to users in config files
 - Allow Horses to Eat Green Apples
+- Update Recipes that use Eggs to allow the use of #forge:eggs Tag
 
 # 1.7.5
 - Fix Incorrect English Language key for Granny Smith Advancement